backend/requirements.txt
anthropic
python-dotenv
playwright
Pillow
fastapi
uvicorn
websockets
pydantic
requests
termcolor
pyfiglet

# For the stream
aiortc
av
numpy